MB-49745 Disallow loading sample buckets on mixed versions
When running mixed versions and attempting to load the sample buckets
we'll now return an error:

$ xcurl localhost:9001/sampleBuckets/install -d '["beer-sample"]'
This http API endpoint isn't supported in mixed version clusters

Note this enforcement starts with Neo. If the sample bucket is loaded
from a pre-Neo node then the attempt will still occur (and may or may
not succeed).

Change-Id: Ieb508114e4ce8edeb53ce1d740bb04a19b77462a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/166550
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Meni Hillel <meni.hillel@couchbase.com>