Add checks to make sure only JSON-ifiable terms are added as event logs
Call ejson:encode/1 to make sure an event log is encodable else we could accept
an event log which cann't be encoded and it will nuke /events endpoint.

There is no specific bug that prompted this change - but it's a good fix
to make sure, we don't loose the other event logs due to a single
bad event log.

Change-Id: I92820830d2a6bd241a7cee6015140e59b452cfa9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/167153
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Meni Hillel <meni.hillel@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>