MB-49795 Better message when multi-node kv autofailover couldn't complete
In a situation where the number of down kv nodes is greater than maximum
autofailover count and we are duty bound to preserve durable mutations,
we cannot perform auto-failover of a subset of kv nodes limited to maximum
autofailover count. This is due to the fact that we cannot contact the
nodes that are not going to be failed over but are down to determine
their replication status which is key information required to promote
max replica.

Change-Id: Ia3f4313d87e0b01fcaaceb043a730aceadfa30c5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/167431
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>