MB-49895 Tag raw_url for redaction
When a raw_url is included in an audit log it is tagged for redaction
because it may contain sensitive information such as a user name.
The reason the entire url is tagged is that, from the ns_server POV,
just an arbitrary string.

Change-Id: I16021f5fb9518d8b203a66e9b3b9873258c8cd32
Reviewed-on: https://review.couchbase.org/c/ns_server/+/167971
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>