MB-50132 Fix menelaus_web_cache invalidation predicate for ... bucketsStreaming
The current invalidation predicate checks if OldUpdateId > UpdateId -
which results in an incorrect check when OldUpdateId is 'undefined'.

Because of which, all changes to the bucket that happen before the cache
entry expires (in 1 second) for the first time, are effectively not
reflected in the bucketsStreaming endpoint.

Change-Id: Ib50cbdf3aecba42a98cc7feaa3d300a50105b94f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/168212
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>