MB-49646 Protect ns_config from changes by removed nodes.
To achieve that ns_config_rep keeps the list of known
nodes (nodes_wanted) and the corresponding chronicle revision. Since
all nodes will observe changes to nodes_wanted at varying moments of
time, the revision is used to help guide the decision regarding
whether incoming changes should be accepted.

Ignoring quorum failovers, the logic is the following:

1. If node A gets replicated to by node B that has a lower revision,
the corresponding changes will only be accepted if node B is in the
topology known by node A.

2. If, on the other hand, node A's revision is behind node B's
revision, node A may not consult its topology (since it's out of
date). So node A will always accept mutations in such case.

Dealing with quorum failovers is a bit more complicated because
revisions may not be comparable. This is possible in two cases:

1. Node A and node B are both part of the surviving partition, but
node B is ahead of node A. Ultimately, in this case we would like to
let the changes go through.

2. Node A is part of the surviving partition, while node B is part of
the partition that was failed over. In this case no changes from node
B should be accepted.

Distinguishing between these two situations requires more metadata to
be sent coupled with all replicated changes (and also greater
complexity), so I decided against this route.

Instead, any changes will be conservatively rejected if they come from
a node with an incompatible revision. But once the recipient node
learns about the quorum failover, it will resync with all surviving
nodes.

Change-Id: Ic2d20156b44fc862d18cfd85cf7ed9a467597f45
Reviewed-on: https://review.couchbase.org/c/ns_server/+/168340
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>