MB-50188 Handle create_bucket failure
When the connection from ns_server to memcached completes we try to
ensure the bucket. There is a window (sometimes very long) where the
bucket_select fails with enoent and when we try to create the bucket it
fails with eexists. The reason is memcached returns the enoent for a
bucket that exists yet hasn't completed coming online.

This change captures the error from the bucket create and retries after
waiting a bit.

Change-Id: I64bf85c340fe7a14c7a94896e3d878295413b666
Reviewed-on: https://review.couchbase.org/c/ns_server/+/168345
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>