MB-50256: create ObserveSeqno privilege
Created an ObserveSeqno privilege that operates at the bucket level so
that the existing MetaRead privilege can continue to operate at the
scope/collections level while exposing the required information for
the elastic-search adaptor to be able to operate. Otherwise it would
require MetaRead at the entire bucket which could complicate
multi-tenancy use-cases. This is the ns_server side of the work, but
kv will be required to understand this new privilege before it can be
used.

This is tied to the KV commit (starting with):
https://review.couchbase.org/c/kv_engine/+/168612

Change-Id: If365393acce9657ecaa59a7aaea691304d0c2625
Reviewed-on: https://review.couchbase.org/c/ns_server/+/168459
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>