MB-50433: Add event log only when the Audit config has changed
1) Add event log only when the audit config has changed - previously the
intent was to capture any change done by the user; so there wasn't any
logic to check if the NewSettings and OldSettings were different or not.

2) The Settings received via the UI don't include all the keys present
in the audit config (ex: sync). A couple of new helper functions are
added to retrive the OldSettings and NewSettings that are printed in the
audit_cfg_changed event log.

Change-Id: I12592ff83f7d01ad050607a88e5cb5b2a1564ebc
Reviewed-on: https://review.couchbase.org/c/ns_server/+/168773
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>