MB-50406: User creation and deletion uses too much memory
The user storage subsystem uses multiple tables to manager users
throughout the cluster. These systems include a DETS table for
persistent storage and an ETS table that is used as a cache. When we
start up we convert the DETS table into the ETS table, which is then
used for querying.

This means we have a 1-to-1 copy of the table in two places. Although
the max size of a DETS table is 2gb, the ETS table can also grow very
large with a large number of users created. The tombstones are also
kept in memory, which means that you can also increase the used memory
by creating and deleting users. That means a customer could only have
100 active users, at any given time, but by cycling through random
usernames and deleting them, the memory will grow indefinitely.

The solution, taken in this patch, is to only keep active users in the
ETS table and relegate the tombstones to only the DETS table. This
means we would only hold those 100 users that are currently active in
memory, instead of holding all the deleted users as well. There are a
few other mechanisms, such as select, that needed to be modified.
There is an "all_docs" function that requires us to pull all users
from the DETS table, while normal selects can happen over the
in-memory ETS table. Individual user lookups will go to the ETS table
first, and then to the DETS table if it's not found in the ETS table.

Change-Id: Ib48749d105040ab1c9e464d56e53851c05a3d189
Reviewed-on: https://review.couchbase.org/c/ns_server/+/169514
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>