MB-49868 Sample buckets have 0 replicas on single node
If a sample bucket is loaded on a single node cluster then the number
of replicas will be zero. This is to allow running of transactions
and without this change will fail with durability impossible errors.

This change is dependent on cbimport providing support for specifying
the number of replicas:

https://review.couchbase.org/c/backup/+/169183

Change-Id: Ifaa47c730aa190718862f0f30655f4404e1ff891
Reviewed-on: https://review.couchbase.org/c/ns_server/+/169184
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>