MB-50442 release control back to dcp_sup from newly created
...dcp_replicator right after the local connection to memcached is
created but before the remote connection is established. make sure
that during the creation of the remote connections exits are processed
and hanged connect can be aborted

this is needed to make sure that if the autofailover aborts the rebalance
the rebalance artifacts can be quickly cleared without waiting for
the timeout if the remote memcached port is non responsive. the clearing
of the rebalance artifacts must happen before the query_vbuckets call
is returned, so it being too slow might cause timeout of query_vbuckets
call and therefore failover to fail.

Change-Id: Ib855fba3407206c4b815672cc974f9b3c4b9899e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/169185
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>