MB-50509 Allow resetting compaction sizes
When the Database Fragmentation Size or View Fragmentation Size is
changed from their default value, 'undefined', there was no way to
change it back to 'undefined'.

This change allows specifying a size of zero which is then converted
to 'undefined', thus allowing it to be reset to its default value.

Change-Id: Ia4656b6fc4e07baf2843cde89b14049e2a2caf2c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/169527
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>