MB-50368: Redact ldap cfg keys in event log
1. prepare_ldap_settings/1 transforms the keys in the ldap config to binaries
and the keys for redaction we were sending were atoms.

The logic got changed during a re-factor; reverting the logic to the
older logic of redacting the keys first and then running
prepare_ldap_settings/1.

2. Add extra function closure to event_log:maybe_add_log_settings_changed/4
which takes a TransformFun which jsonifies (and/or redacts etc) the
Settings.

Change-Id: I2f1ce27de8c6bfea13816ad241b5b557f8462cee
Reviewed-on: https://review.couchbase.org/c/ns_server/+/169633
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>