MB-49876: AddNode fails due to exception on ns_couchdb node
This fix addresses a race between ns_couchdb node and rename of
ns_server node. If the notification of the name change arrives at the
ns_couchdb node late enough, enough children processes under
ns_couchdb_sup can restart to satisfy the max_restart_intensity of the
supervsior for it to just give up, hence causing the ns_couchdb to be
shut down permanently by the supervisor. This fix hardens the
requirment for a shutdown of ns_couchdb in such a scenerio.

Change-Id: I8ba2231ca2eddd8a154c805e7fa70a1beab9a20e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/169789
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>