MB-50711: Include user identity in framing info extras object
This fix will include the user identity in the 'FrameInfo' object of
the 'flexible framing extras' described in the Couchbase Binary
Protocol. This fix does this for the 'Subdoc multi lookup` OpCode type
as the lack of the identity information for this OpCode was the cause
of the missing identity info in the audit logs.

This Couchbae Binary Protocol is described here:
http://src.couchbase.org/source/xref/trunk/kv_engine/docs/BinaryProtocol.md

Change-Id: I36ea75668ec5322fadf35650d18c8e4665822660
Reviewed-on: https://review.couchbase.org/c/ns_server/+/170040
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>