MB-50794 Don't tag items in audit logs
Prior to this change the read_doc, mutate_doc, auth_failure, and
rbac_info_retrieved audit log entries contained tagged (to be redacted)
information. As audit.logs are not collected files there is no need to
tag the items.

The info that was previosly tagged in the audit logs is now tagged prior
to getting emitted into the logs (e.g. debug.log) that are collected and
possibly redacted.

Change-Id: Iddcdf7e91994f42c9bf6b789b864ccba1755564a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/170052
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>