MB-50801 Log topology-aware service needing rebalance
There are two paths where the switching of the compatibility mode is
done. The first is at the end of rebalance. The second is during the
periodic check by the ns_janitor_server. Only for the first case will
we check for new topology-aware services and do the service upgrade.
In this case n1ql, an existing service, was made topology-aware in the
Neo release.

Trying to upgrade the services by the ns_janitor_server would require
asking the ns_orchestrator to do the service upgrade. This itself would
introduce a new state in the state machine and might not be possible if
the orchestrator was not idle (which means the ns_janitor_service would
have to remember to periodically poke the orchestrator to do the service
upgrade).

A rebalance will upgrade the service. Changes to just perform the
service upgrade portion of a rebalance are significant. As a result a
new "User Log" will be emitted providing detail as to why the rebalance
is needed.

Change-Id: I44797c1bbf93350a573f313fbfc7aecb73ec2326
Reviewed-on: https://review.couchbase.org/c/ns_server/+/170415
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>