MB-50708 discontinue support of the following memcached permissions:
...MetaRead, XattrRead, ReadSeqno, XattrWrite

MetaRead is replaced with a check for Read against the same key.
So if a role has Read on the same collection, scope or bucket,
MetaRead can be simply removed from that role.

XattrRead is being removed and replaced with a Read check against
the same key.

ReadSeqno: The check on this privilege is replaced by a check
for Read on any collection or scope in the bucket.

XattrWrite is replaced by a check for Write against the same key

Change-Id: I574f242951b3b624e985c9c57f56005af14e5801
Reviewed-on: https://review.couchbase.org/c/ns_server/+/170924
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>