MB-50868: Filter out failed nodes from bucket server config list
During failover janitor cleanup, if the serverlist of a bucket is only
currently tied to a single node at the time of failover, the janitor
was incorrectly setting the server list to also include the node being
failed over since it uses the active KV nodes to set the initial list,
and Node status doesn't become inactive until janitor cleanup is done.
This fix addresses that by setting a correct list that accounts for the
failed node in this case.

Change-Id: I8ddf4d57f004c3dbbae5ef14b46c617efb0a57de
Reviewed-on: https://review.couchbase.org/c/ns_server/+/170697
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>