MB-50708 support meta read capability for "XDCR Inbound" role on
...mixed clusters

Neo node is not checking for cluster.bucket[name].data.meta!read
permission anymore to allow read access to meta in memcached.
Instead the read access to meta is granted if there is a read access
to docs. Therefore not to lose read access to meta on Neo nodes added
to pre Neo cluster we need to add the following permission to pre Neo
"XDCR Inbound" roles: cluster.bucket[name].data.docs!read

The "or read any data" part of the description of the role thus becomes
incorrect, but I decided to leave it as is not to create confusion
of having different role descriptions on different nodes of the cluster

Change-Id: I82a13c3f7207df9d39f769d2d69c9bf89e85f5c4
Reviewed-on: https://review.couchbase.org/c/ns_server/+/170925
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>