MB-51197: Add tls status of node to dump-guts/cbcollect
Adds tls support to dump-guts, and cbcollect so that it can be used to
alter the behavior of the projector collector task and make the
correct requests to https instead of http when cluster encryption is
enabled. Also makes this information generally available for other
tasks to use in cbcollect, as needed.

Change-Id: I5340780419a6e51bff9bbb505fed71dde99c2f3e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/171549
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>