MB-51197: add curl '-k' (allow-insecure) option by default to get
Allow insecure connections that don't require the proper certificate
for curl to connect to self-signed or unknown certificates on
cbcollect. Used primarily for collecting projector dumps but is also
added by default for get requests and can be expanded to other
requests in the future, if needed.

Change-Id: I76fea50e30f80436278cb4bd996325cd0045037e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/171694
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>