MB-51125: Use snapshot for bucket config for consistency
Originally was looking up the information again at the point of the
basic_stats call which meant the bucket list and the current existing
buckets could differ if we were in-between bucket create/delete
operations. In the case of this issue, it was a bucket being deleted
that was available at the time of the iteration but not the sub-call
that actually looked up the basic_stats information. This
inconsistency is fixed by using a snapshot in all places instead of
making additional lookups in the sub-functions.

Change-Id: I1a395888cbf87f5fff96a48e075e932e3c81b6fc
Reviewed-on: https://review.couchbase.org/c/ns_server/+/171808
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>