MB-49705 [BP] Don't leak sensitive information (complete_join)
Prior to this change a crash while doing an engage_cluster could leak
sensitive information such as PRIVATE KEYs. We know wrap the list of
arguments used by engage_cluster in a function and pass the function
instead.

This is a backport of MB-48222.

Change-Id: I30ac0c0f4a6c5542519619328dc21981bdb81de6
Reviewed-on: https://review.couchbase.org/c/ns_server/+/172373
Well-Formed: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>