MB-51620 Validate prometheus snapshot directory
When cbcollect uses the _createStatsSnapshot endpoint the results may be
an error message, instead of the expected directory path containing the
stats snapshot. Prior to this change the error message was treated as a
directory and we failed to get stats. With this change we validate the
results to be an existing directory and if it isn't the case then the
stats are obtained directly from disk.

Change-Id: I63fa7552b34ee54b60baa0e70e39f3d255e7e997
Reviewed-on: https://review.couchbase.org/c/ns_server/+/173054
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>