MB-51771 MB-51776: Add clientCertAuth to /addNode and /doJoinCluster
When node is joining the cluster admin should be able to choose
client certs to be used for authentication. Before this change user
always had to provide username and password for the remote node
(or cluster). If the remote node is configured to use client cert for
authentication, there should be a way to use it for node addition.

When clientCertAuth=true is passed in /addNode it means that
the-cluster-node will try to use client certificate for auth, when
calling /engageCluster and /completeJoin.

When clientCertAuth=true is passed in /doJoinCluster it means that
the-node-to-be-added will try to use client certificate for auth, when
calling /addNode against the-cluster-node.
Note that the-node-to-be-added will set clientCertAuth=true in
the /addNode call only if client-cert-auth usage is configured to be
'mandatory' at the-node-to-be-added.

Note that the client certificate is not passed in neither of the
calls. Nodes use their internal (generated or uploaded) client
certs for that.

Change-Id: I441970d81a52a81e6f88a15f956a97af9c36342d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/173592
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>