MB-51808 Return correct mem info for uninitialized node
Prior to the use of prometheus by ns_server the mem_total and mem_free
system stats would not be available (aka undefined) while the node is
uninitialized. With the introduction of prometheus the stats are no
longer undefined but may be zero. This change accounts for that case.

Change-Id: I141a0bd2d5e4475cb3d1e2b116c2edfb780e2074
Reviewed-on: https://review.couchbase.org/c/ns_server/+/173654
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>