MB-51814 Return number of vbuckets
The buckets related endpoints will now return the number of vbuckets for
the bucket. Is is in preparation for having different vbucket counts for
each bucket. The FTS, Index, and eventing services currently are
started with the number of vbuckets as startup arguments. These services
will need to switch to using a rest endpoint to get the number of
vbuckets.

$ xcurl localhost:9000/pools/default/b/travel-sample | jq | grep numV
  "numVBuckets": 16,

Change-Id: Id1eb2cef52596d9ce4c002c60a6da6aee4bc23b9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/173769
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>