MB-51665 Don't raise false diskspace worker stuck
We periodically run a check of diskspace and that check can possibly
hang and so we have logic to detect if it has hung. In addition we might
be running on a laptop which has been closed and then opened after some
time. We currently cannot tell if the lack of a periodic diskspace check
is due to it being wedged or due to the time jumping.

We now will not declare the diskspace check to be wedged until it has
not updated the diskspace info for three times the periodic interval. If we
wake up from a laptop being reopened then enough time may have elapsed
such that we mark the diskspace check as pending stale and when the
second interval elapses and the check doesn't occur we assume it's wedged.
We allow three misses to account for possible races between the process
that updates the last checked time and the process doing the check.

Change-Id: I3e227f8588000eee5a3fc5d0afcf62fb18c21f3a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/174401
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>