MB-51949 - Make workbench sidebar scrollable
For both the Analytics and Query workbenches, the right sidebar used
to be scrollable when it was too tall to fit on the screen. Recent
code cleanup broke this behavior, making the whole window scroll. This
css change brings back scrolling for the right sidebar in both
workbenches.

Change-Id: Idb604ca56a5f741c5a738d671902ccf247114261
Reviewed-on: https://review.couchbase.org/c/ns_server/+/175792
Tested-by: Eben Haber <eben@couchbase.com>
Tested-by: Rob Ashcom <rob.ashcom@gmail.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Rob Ashcom <rob.ashcom@gmail.com>
Reviewed-on: https://review.couchbase.org/c/ns_server/+/176138
Reviewed-by: Eben Haber <eben@couchbase.com>
Well-Formed: Restriction Checker