MB-52464 UID for initial bucket manifest shouldn't be zero
The initial bucket manifest UID for a bucket should not be zero unless
it only contains the _default scope and _default collection as kv relies
on this.

For serverless configs (where the "_system" scope is included) the
initial bucket manifest UID is now 1 to not violate kv's assumption.

Change-Id: If5fdcd6608fb0dc3fc3c665de00017d8914ee8bb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/175839
Reviewed-by: James Lee <james.lee@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>