Refactor users migrations code and some new tests
The main purpose of the refactoring is to prepare the code for adding
migrations for auth records (current implementation only allows
migrations of users and groups). Also modify migration tests for the
same reason.

Also a couple of new tests are added by this change for already
existing migrations:
 - upgrade of groups roles test
 - add uuid for local users test

Change-Id: I57cdc00a09f0dfff7617f4564ce7357d22a884ad
Reviewed-on: https://review.couchbase.org/c/ns_server/+/175848
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>