MB-52258, MB-52406: Introduce file-based switch for profile(s)
Allow setting configuration profile by setting the contents of a file
located at /etc/couchbase.d/config_profile. If that file contains
"serverless" we will load the serverless profile from our products
'etc' directory. If the file is not found, or it contains "default"
the default profile will be used. If another name is provided, but no
file is found to associate with it, we just fallback to default
profile.

Change-Id: Ife7be299cebc73f5cb8891be06c9231565c22a36
Reviewed-on: https://review.couchbase.org/c/ns_server/+/175957
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>