MB-51986: Set serverless 'max_vbuckets' default to 48
During testing for serverless we varied the number of vbuckets from 4 to
1024 and found that 48 vbuckets hit a nice sweet spot that provided
enough throughput to get the performance we required. But also helped to
reduce the amount of CPU cycles and memory we used at idle. 48 also has
the nice property of being a divisible by 2 and 3.

Change-Id: I5fc10740a8200522a8a38ddfb8288ff3400cd463
Reviewed-on: https://review.couchbase.org/c/ns_server/+/176526
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>