MB-52197: Introducing new cbauth GetUserBuckets API
The API returns which buckets the user can access. The user might only
have permissions for a specific collection in the bucket but is deemed
able to access to the bucket.
API is currently required for query so that given a user name, they know
what bucket we need to use as tenant for metering and throttling purposes.

Corresponding cbauth change,
https://review.couchbase.org/c/cbauth/+/176997

Change-Id: I75609f121d8370522abf751c0477047c8148340a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/176996
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>