MB-23768: Call menelaus_roles:validate_roles when validating...
... roles  that come from http endpoints

Basically don't start addition of groups or users if
the menelaus_roles:validate_roles check doesn't pass.
It is checked in menelaus_users but for the case of backup
restore we don't want to create half of all the users and
then find out that some role is invalid. We need to know
it before we create any user or group.

Change-Id: Ia35f4251f8dd4a30c58c7eb54b949d9fa55d3458
Reviewed-on: https://review.couchbase.org/c/ns_server/+/177860
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>