MB-23768: Move security_roles_access and ldap_access checks...
... to validators

This is a preparation for adding backup restore code.
Trying to minimize the probability of user backup restore being
interrupted by something that can be checked beforehand.
Ideally, if validation succeeds, user addition should succeed
as well. It's hard to guarantee that but it's best efford.

Change-Id: I189bc1417952049817067c343f9070b1520f601d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/177863
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>