MB-53077: Add RAM usage alerts (part 2)
This patch deals with pop up alerts.
They didn't exist in Mad-Hatter so we didn't need
to deal with it there.

Strategy is the same as with email alerts.
We save info about the memory alert in a separate
ns_config key (pop_up_alerts), in order to avoid
backward incompatibility.

Change-Id: Idf610f50774d518510e691bae5c83a79c5d35997
Reviewed-on: https://review.couchbase.org/c/ns_server/+/178008
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>