MB-53323: consider keep nodes when placing buckets in rebalance
Prior to this patch, bucket placement only considered active server
group nodes when placing buckets. This leads to issues where a failed
over node is being recovered and will be active as soon as the rebalance
starts. This patch considers nodes that will be kept as part of the
rebalance in addition to active nodes in server groups when computing
the bucket placement.

There remains the issue that it is impossible to remove all the nodes in
a given server group. We currenlty require that the nodes to be removed
are moved to a different server group, the original server group is
deleted and then the nodes are remove and rebalance is run. But this we
can think about.

Change-Id: I58aca2d9c126f62e023e87895558efd2d7749448
Reviewed-on: https://review.couchbase.org/c/ns_server/+/178677
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>