MB-53193: Reenable autofailover popup alerts
when which were disabled by 7.0 upgrade by mistake.

Yes, we also can mistakenly reenable alerts that were
intentionally disabled by users, but it's better to have alerts
enabled by mistake than alerts disabled by mistake.

This is an upgrade to elixir which is basically the same as
upgrade to 7.1 so we can repair the popup config for customers
who are already on 7.1.x.

Change-Id: I44e5072c4f9ec9c556ecffcadb39118c6c547ea7
Reviewed-on: https://review.couchbase.org/c/ns_server/+/178728
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>