MB-53423 Adjust bucket maximums for _system scope
For buckets with _system scope (and the associated "special"
collections) the initial number of scopes/collections in the manifest
is set to zero even though there's "_default", and possibly "_system"
scopes as well as special collections. This is done to avoid possibly
confusing errors e.g. "Maximum number of collections (103) for a
bucket has been reached".

Change-Id: I89420842748a348fd5aa150b98b09ca1098459bb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/178861
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>