MB-51516: Don't clamp in_docs estimate to current checkpoint items
Currently ns_rebalance_observer clamps the in_docs estimates of a
rebalance to the number of items in the vBucket's currently open
Checkpoint. This results in the incoming docs of a node not
participating in a swap rebalance (i.e. A of [A, B, C] -> [A, B, D])
being non-zero when no documents logically need moving and no documents
are in fact moved.

Clamping to the current checkpoint items only makes the count more
accurate for full eviction Buckets with a large queue of unpersisted
items. It does so at the expensive of less accuracy for at rest Buckets
and value eviction Buckets which I don't believe is a worthwhile
tradeoff.

Remove the clamp to the number of checkpoint items and instead allow the
in_docs estimate delta to be as low as 0.

Change-Id: Ib2d83728596f1872931f7b934926473a4648187f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/178944
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>