MB-53478: Fix saving anonymous functions to disk
Instead of saving a callback to disk and reloading it, we replace it
while we are loading them to include a dummy value instead of the
closure that was saved originally. This solves the issue where we get
stuck with items in the queue with bad continuations between ns_audit
restarts. On new code, we don't save the closure at all, but instead
pre-convert the synchronous entries into asynchronous ones, so that we
don't attempt to reply to a request that has already failed.

Change-Id: If42c622317c7c93ad59d2a9a931e639869e43db0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/179138
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>