MB-60839: [BP] Fix saving anonymous functions
It was difficult to extricate one change from the other and the addition
of MB-52152 is a positive one with low risk. It also makes merging
forward much simpler as the changes diverge less from master.

 * MB-53478: Fix saving anonymous functions to disk

Instead of saving a callback to disk and reloading it, we replace it
while we are loading them to include a dummy value instead of the
closure that was saved originally. This solves the issue where we get
stuck with items in the queue with bad continuations between ns_audit
restarts. On new code, we don't save the closure at all, but instead
pre-convert the synchronous entries into asynchronous ones, so that we
don't attempt to reply to a request that has already failed.

 * MB-52152: Don't send audit if auditing is disabled

If auditing is disabled, we will continue to log the audit to the
debug.log for aid in triage but won't queue it to be sent to memcached.

Change-Id: Ie1c48de2d6b0a1be93d421b40e6a62de5273f293
Reviewed-on: https://review.couchbase.org/c/ns_server/+/207366
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>